+++ packages/ui/src/elements/BulkUpload/AddFilesView/index.scss
@@ -7,9 +7,22 @@
     height: 100%;
     padding: calc(var(--base) * 2) var(--gutter-h);
   }
-
+  
   .dropzone {
     flex-direction: column;
     justify-content: center;
+    display: flex;
+    gap: var(--base);
+    background-color: var(--theme-elevation-50);
+
+    p {
+      margin: 0;
+    }
+  }
+
+  &__dragAndDropText {
+    margin: 0;
+    text-transform: lowercase;
+    align-self: center;
   }
 }
