+++ packages/ui/src/elements/Dropzone/index.tsx
@@ -1,8 +1,6 @@
 'use client'
 import React from 'react'
 
-import { useTranslation } from '../../providers/Translation/index.js'
-import { Button } from '../Button/index.js'
 import './index.scss'
 
 const handleDragOver = (e: DragEvent) => {
@@ -10,30 +8,25 @@ const handleDragOver = (e: DragEvent) => {
   e.stopPropagation()
 }
 
-const baseClass = 'dropzone'
+const baseClass = 'dropzone-shell'
 
 export type Props = {
   readonly children?: React.ReactNode
   readonly className?: string
-  readonly mimeTypes?: string[]
+  readonly dropzoneStyle?: 'default' | 'none'
   readonly multipleFiles?: boolean
   readonly onChange: (e: FileList) => void
-  readonly onPasteUrlClick?: () => void
 }
 
-export const Dropzone: React.FC<Props> = ({
+export function Dropzone({
   children,
   className,
-  mimeTypes,
+  dropzoneStyle = 'default',
   multipleFiles,
   onChange,
-  onPasteUrlClick,
-}) => {
+}: Props) {
   const dropRef = React.useRef<HTMLDivElement>(null)
   const [dragging, setDragging] = React.useState(false)
-  const inputRef = React.useRef(null)
-
-  const { t } = useTranslation()
 
   const addFiles = React.useCallback(
     (files: FileList) => {
@@ -88,15 +81,6 @@ export const Dropzone: React.FC<Props> = ({
     [addFiles],
   )
 
-  const handleFileSelection = React.useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      if (e.target.files && e.target.files.length > 0) {
-        addFiles(e.target.files)
-      }
-    },
-    [addFiles],
-  )
-
   React.useEffect(() => {
     const div = dropRef.current
 
@@ -119,45 +103,18 @@ export const Dropzone: React.FC<Props> = ({
     return () => null
   }, [handleDragEnter, handleDragLeave, handleDrop, handlePaste])
 
-  const classes = [baseClass, className, dragging ? 'dragging' : ''].filter(Boolean).join(' ')
+  const classes = [
+    baseClass,
+    className,
+    dragging ? 'dragging' : '',
+    `dropzoneStyle--${dropzoneStyle}`,
+  ]
+    .filter(Boolean)
+    .join(' ')
 
   return (
     <div className={classes} ref={dropRef}>
-      <Button
-        buttonStyle="secondary"
-        className={`${baseClass}__file-button`}
-        onClick={() => {
-          inputRef.current.click()
-        }}
-        size="medium"
-      >
-        {t('upload:selectFile')}
-      </Button>
-      {typeof onPasteUrlClick === 'function' && (
-        <Button
-          buttonStyle="secondary"
-          className={`${baseClass}__file-button`}
-          onClick={onPasteUrlClick}
-          size="medium"
-        >
-          {t('upload:pasteURL')}
-        </Button>
-      )}
-      <input
-        accept={mimeTypes?.join(',')}
-        aria-hidden="true"
-        className={`${baseClass}__hidden-input`}
-        multiple={multipleFiles}
-        onChange={handleFileSelection}
-        ref={inputRef}
-        type="file"
-      />
-
       {children}
-
-      <p className={`${baseClass}__label`}>
-        {t('general:or')} {t('upload:dragAndDrop')}
-      </p>
     </div>
   )
 }
