--- packages/ui/src/elements/DropzoneShell/index.scss
@@ -1,34 +1,0 @@
-@import '../../scss/styles.scss';
-
-.dropzone {
-  position: relative;
-  display: flex;
-  align-items: center;
-  padding: calc(var(--base) / 3) calc(var(--base) / 2);
-  background: var(--theme-elevation-50);
-  border: 1px dotted var(--theme-elevation-400);
-  border-radius: var(--style-radius-s);
-  height: 100%;
-  width: 100%;
-  box-shadow: 0 0 0 0 transparent;
-  transition: all 100ms cubic-bezier(0, 0.2, 0.2, 1);
-
-  &.dragging {
-    border-color: var(--theme-success-500);
-    background: var(--theme-success-150);
-    @include shadow-m;
-
-    * {
-      pointer-events: none;
-    }
-  }
-
-  @include mid-break {
-    display: block;
-    text-align: center;
-  }
-
-  &.dropZoneStyle--none {
-    all: unset;
-  }
-}
