--- packages/ui/src/elements/FileDetails/UploadDocRelationship/index.scss
@@ -1,36 +1,0 @@
-.uploadDocRelationshipContent {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-
-  &__imageAndDetails {
-    display: flex;
-    gap: calc(var(--base) / 2);
-    align-items: center;
-  }
-
-  &__thumbnail {
-    align-self: center;
-    border-radius: var(--style-radius-s);
-  }
-
-  &__details {
-    display: flex;
-    flex-direction: column;
-    gap: 0;
-  }
-
-  &__title {
-    margin: 0;
-  }
-
-  &__meta {
-    margin: 0;
-    color: var(--theme-elevation-500);
-  }
-
-  .btn {
-    margin: 0;
-  }
-}
