+++ packages/ui/src/fields/Upload/UploadCard/index.scss
@@ -1,4 +1,4 @@
-.upload-field-row-card {
+.upload-field-card {
   background: var(--theme-elevation-50);
   border: 1px solid var(--theme-border-color);
   border-radius: var(--style-radius-s);
@@ -8,7 +8,7 @@
   gap: calc(var(--base) / 2);
 
   &--size-medium {
-    padding: calc(var(--base) / 2);
+    padding: calc(var(--base) * .5);
 
     .thumbnail {
       width: 40px;
