+++ packages/ui/src/fields/Upload/UploadCard/index.tsx
@@ -2,14 +2,14 @@ import React from 'react'
 
 import './index.scss'
 
-const baseClass = 'upload-field-row-card'
+const baseClass = 'upload-field-card'
 
 type Props = {
   readonly children: React.ReactNode
   readonly className?: string
   readonly size?: 'medium' | 'small'
 }
-export function RowCard({ children, className, size = 'medium' }: Props) {
+export function UploadCard({ children, className, size = 'medium' }: Props) {
   return (
     <div className={[baseClass, className, `${baseClass}--size-${size}`].filter(Boolean).join(' ')}>
       {children}
